# backend/internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "weflow swagger api documention.",
        "title": "weflow",
        "contact": {
            "name": "hertz-contrib",
            "url": "https://github.com/hertz-contrib"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/weflow",
    "paths": {
...
        "/org/list": {
            "post": {
                "description": "查询组织列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户中心"
                ],
                "summary": "查询组织列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "OrgQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.OrgQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.OrgInfoResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/role/list": {
            "post": {
                "description": "查询角色列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户中心"
                ],
                "summary": "查询角色列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "RoleQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.RoleQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.RoleInfoResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/list": {
            "post": {
                "description": "查询用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户中心"
                ],
                "summary": "查询用户列表",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "UserQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UserQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.UserInfoResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/usertask/agree": {
            "post": {
                "description": "同意用户任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户任务"
                ],
                "summary": "同意用户任务",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "UserTaskAgreeVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UserTaskAgreeVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/usertask/disagree": {
            "post": {
                "description": "不同意用户任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户任务"
                ],
                "summary": "不同意用户任务",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "UserTaskDisagreeVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UserTaskDisagreeVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/usertask/done": {
            "post": {
                "description": "获取已办用户任务列表（已处理）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户任务"
                ],
                "summary": "获取已办用户任务列表（已处理）",
                "parameters": [
                    {
                        "description": "已处理的请求参数",
                        "name": "UserTaskQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UserTaskQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.UserTaskResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/usertask/received": {
            "post": {
                "description": "获取用户任务列表（我收到的）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户任务"
                ],
                "summary": "获取用户任务列表（我收到的）",
                "parameters": [
                    {
                        "description": "我收到的的请求参数",
                        "name": "UserTaskQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UserTaskQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.UserTaskResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/usertask/save": {
            "post": {
                "description": "保存用户任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户任务"
                ],
                "summary": "保存用户任务",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "UserTaskSaveVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UserTaskSaveVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "$ref": "#/definitions/base.Response"
                        }
                    }
                }
            }
        },
        "/usertask/todo": {
            "post": {
                "description": "获取待办用户任务列表（待处理）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户任务"
                ],
                "summary": "获取待办用户任务列表（待处理）",
                "parameters": [
                    {
                        "description": "待处理的请求参数",
                        "name": "UserTaskQueryVO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UserTaskQueryVO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/base.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/bo.UserTaskTodoResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "base.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "0:成功，其他：失败",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
...
        "bo.OrgInfoResult": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "createUser": {
                    "description": "创建人",
                    "type": "string"
                },
                "id": {
                    "description": "唯一id",
                    "type": "integer"
                },
                "orgID": {
                    "description": "组id",
                    "type": "string"
                },
                "orgName": {
                    "description": "组名称",
                    "type": "string"
                },
                "parentID": {
                    "description": "组父id",
                    "type": "string"
                },
                "remark": {
                    "description": "描述",
                    "type": "string"
                },
                "status": {
                    "description": "状态【1：未启用；2：已启用；3：锁定；】",
                    "type": "integer"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "updateUser": {
                    "description": "更新人",
                    "type": "string"
                }
            }
        },
        "bo.RoleInfoResult": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "createUser": {
                    "description": "创建人",
                    "type": "string"
                },
                "id": {
                    "description": "唯一id",
                    "type": "integer"
                },
                "parentID": {
                    "description": "角色父id",
                    "type": "string"
                },
                "remark": {
                    "description": "描述",
                    "type": "string"
                },
                "roleID": {
                    "description": "角色id",
                    "type": "string"
                },
                "roleName": {
                    "description": "角色名称",
                    "type": "string"
                },
                "status": {
                    "description": "状态【1：未启用；2：已启用；3：锁定；】",
                    "type": "integer"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "updateUser": {
                    "description": "更新人",
                    "type": "string"
                }
            }
        },
        "bo.UserInfoResult": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "createUser": {
                    "description": "创建人",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "唯一id",
                    "type": "integer"
                },
                "orgID": {
                    "description": "组织id",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "remark": {
                    "description": "描述",
                    "type": "string"
                },
                "status": {
                    "description": "状态【1：未启用；2：已启用；3：锁定；】",
                    "type": "integer"
                },
                "updateTime": {
                    "description": "更新时间",
                    "type": "string"
                },
                "updateUser": {
                    "description": "更新人",
                    "type": "string"
                },
                "userID": {
                    "description": "用户id",
                    "type": "string"
                },
                "userName": {
                    "description": "用户名称",
                    "type": "string"
                }
            }
        },
        "bo.UserTaskResult": {
            "type": "object",
            "properties": {
                "createUserID": {
                    "description": "创建人id",
                    "type": "string"
                },
                "createUserName": {
                    "description": "创建人名称",
                    "type": "string"
                },
                "endTime": {
                    "description": "结束时间",
                    "type": "string"
                },
                "instStatus": {
                    "description": "任务状态【1：创建中(草稿)；2：进行中； 3：终止； 4：完成； 5：挂起；6：回退】",
                    "type": "integer"
                },
                "instTaskID": {
                    "description": "实例任务id",
                    "type": "string"
                },
                "nodeID": {
                    "description": "节点id",
                    "type": "string"
                },
                "nodeModel": {
                    "description": "节点模型【1：开始节点；2：审批节点；3：办理节点；4：抄送节点；5：自定义节点；6：条件节点；7：分支节点；8：汇聚节点；9：结束节点】",
                    "type": "integer"
                },
                "nodeName": {
                    "description": "节点名称",
                    "type": "string"
                },
                "nodeTaskID": {
                    "description": "节点任务id",
                    "type": "string"
                },
                "parentID": {
                    "description": "父节点id",
                    "type": "string"
                },
                "startTime": {
                    "description": "发起时间",
                    "type": "string"
                },
                "taskName": {
                    "description": "实例任务名称",
                    "type": "string"
                },
                "userTaskID": {
                    "description": "处理人任务id",
                    "type": "string"
                }
            }
        },
...
        "vo.OrgQueryVO": {
            "type": "object",
            "properties": {
                "orgName": {
                    "description": "组织名称",
                    "type": "string",
                    "example": ""
                }
            }
        },
...
        "vo.RoleQueryVO": {
            "type": "object",
            "properties": {
                "roleName": {
                    "description": "角色名称",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.UserQueryVO": {
            "type": "object",
            "properties": {
                "userName": {
                    "description": "用户名称",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.UserTaskAgreeVO": {
            "type": "object",
            "properties": {
                "opinionDesc": {
                    "description": "操作意见",
                    "type": "string",
                    "example": ""
                },
                "params": {
                    "description": "参数",
                    "type": "object",
                    "additionalProperties": {}
                },
                "userTaskID": {
                    "description": "用户任务id",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.UserTaskDisagreeVO": {
            "type": "object",
            "properties": {
                "opinionDesc": {
                    "description": "操作意见",
                    "type": "string",
                    "example": ""
                },
                "userTaskID": {
                    "description": "用户任务id",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.UserTaskQueryVO": {
            "type": "object",
            "properties": {
                "createTimeEnd": {
                    "description": "提交审批时间-结束 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string",
                    "example": ""
                },
                "createTimeStart": {
                    "description": "提交审批时间-开始 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string",
                    "example": ""
                },
                "createUserID": {
                    "description": "创建人id",
                    "type": "string",
                    "example": ""
                },
                "finishTimeEnd": {
                    "description": "完成审批时间-结束 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string",
                    "example": ""
                },
                "finishTimeStart": {
                    "description": "完成审批时间-开始 yyyy-MM-dd HH:mm:ss:SSS",
                    "type": "string",
                    "example": ""
                },
                "instStatus": {
                    "description": "任务状态，零值为全部",
                    "type": "integer",
                    "example": 0
                },
                "modelID": {
                    "description": "模板ID",
                    "type": "string",
                    "example": ""
                },
                "pageNum": {
                    "description": "页码",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer",
                    "example": 30
                },
                "taskName": {
                    "description": "任务名称",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "vo.UserTaskSaveVO": {
            "type": "object",
            "properties": {
                "opinionDesc": {
                    "description": "操作意见",
                    "type": "string",
                    "example": ""
                },
                "params": {
                    "description": "参数",
                    "type": "object",
                    "additionalProperties": {}
                },
                "userTaskID": {
                    "description": "用户任务id",
                    "type": "string",
                    "example": ""
                }
            }
        }
    }
}